+++ packages/node/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Suport for network endpoint config providing the ability to set headers and rpc batch sizes (#334)
 
 ## [5.0.1] - 2024-07-29
 ### Fixed